+++ engine/CMakeLists.txt
@@ -22,6 +22,7 @@ target_link_libraries("${PROJECT_NAME}Engine"
         astre::Version
         astre::Entry
         astre::Type
+        astre::Formatter
         astre::Process
 
     PUBLIC
