+++ engine/modules/CMakeLists.txt
@@ -2,4 +2,5 @@ add_subdirectory(Native)
 add_subdirectory(Version)
 add_subdirectory(Entry)
 add_subdirectory(Type)
+add_subdirectory(Formatter)
 add_subdirectory(Process)
