+++ engine/modules/Native/include/native/windows/win.h
@@ -31,27 +31,3 @@ namespace astre::native
     using window_handle = HWND;
     using process_handle = HANDLE;
 }
-
-template <>
-struct std::formatter<astre::native::window_handle> : std::formatter<std::string> {
-  auto format(astre::native::window_handle handle, format_context& ctx) const {
-    return formatter<string>::format(
-      std::format("window_handle[{}]", (std::stringstream{} << handle).str()), ctx);
-  }
-};
-
-template <>
-struct std::formatter<astre::native::opengl_context_handle> : std::formatter<std::string> {
-  auto format(astre::native::opengl_context_handle handle, format_context& ctx) const {
-    return formatter<string>::format(
-      std::format("opengl_context[{}]", (std::stringstream{} << handle).str()), ctx);
-  }
-};
-
-template <>
-struct std::formatter<astre::native::device_context> : std::formatter<std::string> {
-  auto format(astre::native::device_context dc, format_context& ctx) const {
-    return formatter<string>::format(
-      std::format("device_context[{}]", (std::stringstream{} << dc).str()), ctx);
-  }
-};
