+++ engine/tests/CMakeLists.txt
@@ -9,12 +9,20 @@ include(GoogleTest)
 
 # Add test executable
 add_executable(${PROJECT_NAME}Tests
+    
     "src/unit_tests.cpp"
 
     # Components Tests
     "modules/Native/native_tests.cpp"
     "modules/Type/type_tests.cpp"
+    "modules/Process/process_tests.cpp"
 )
+
+if(WIN32)
+    target_sources(${PROJECT_NAME}Tests
+        PRIVATE
+            "modules/Process/winapi_process_tests.cpp")
+endif()
 
 target_include_directories(
     ${PROJECT_NAME}Tests
