+++ engine/tests/modules/Process/process_tests.cpp
@@ -1,0 +1,4 @@
+#include <gtest/gtest.h>
+#include "process/process.hpp"
+
+using namespace astre::process;
