+++ engine/tests/modules/Process/winapi_process_tests.cpp
@@ -1,0 +1,4 @@
+#include <gtest/gtest.h>
+#include "process/windows/process_windows.hpp"
+
+using namespace astre::process::windows;
